backend/app/translate/requirements.txt
beautifulsoup4==4.13.3
docx==0.2.4
docx2pdf==0.1.8
fitz==0.0.1.dev2
Markdown==3.7
numpy==2.2.2
openai==1.61.1
opencv_python_headless==4.10.0.84
openpyxl==3.1.5
pdf2docx==0.5.8
pdfdeal==1.0.2
pdfkit==1.0.0
Pillow==11.1.0
PyMySQL==1.1.1
pypdf==5.2.0
pytesseract==0.3.13
python-dotenv==1.0.1
python_docx==1.1.2
python_pptx==1.0.2
redis==5.2.1
Requests==2.32.3
Shapely==2.0.7
